refactor: transaction response
