+++ frontend/lib/core/widgets/notification/notification_card.dart
@@ -5,7 +5,7 @@ import 'package:fundflow/core/themes/app_styles.dart';
 import 'package:fundflow/features/home/bloc/category/category_bloc.dart';
 import 'package:fundflow/features/home/bloc/category/category_state.dart';
 import 'package:fundflow/features/home/models/category.dart';
-import 'package:fundflow/features/transaction/model/transaction.dart';
+import 'package:fundflow/features/transaction/model/transaction_reponse.dart';
 
 class NotificationCard extends StatelessWidget {
   final TransactionResponse transaction;
