+++ frontend/lib/features/image_upload/bloc/image_upload/image_upload_state.dart
@@ -1,4 +1,4 @@
-import 'package:fundflow/features/transaction/model/transaction.dart';
+import 'package:fundflow/features/transaction/model/transaction_reponse.dart';
 import 'package:image_picker/image_picker.dart';
 
 abstract class ImageState {}
